TSO_Serial_Save_Rec  - Inicializado os ArraysLitst de task e task_file na classe interna From_To
WS_SO_Serial_Save
  - No De Para, add verificação se o os array lista de task e task_file são nulll, se sim, não tenta executar loop de atualização.
  - No loop de Task, adicionado select que busca os daddos da task a ser atualizada, para manter a informação do task_seq_oper.

SM_SO_Service_Exec_Task_Sql_005
 - Modificado tipo das varaivies exec_tmp e task_tmp para long.

Act028
  - Devido a mudança na Sql SM_SO_Service_Exec_Task_Sql_005, modificado o cast de Integer.parseInt, para Long.parseLong no parametro de exec_temp e task_temp

+++ app/src/main/java/com/namoadigital/prj001/sql/SM_SO_Service_Exec_Task_Sql_005.java
@@ -18,10 +18,10 @@ public class SM_SO_Service_Exec_Task_Sql_005 implements Specification {
     private int category_price_code;
     private int service_code;
     private int service_seq;
-    private int exec_tmp;
-    private int task_tmp;
+    private long exec_tmp;
+    private long task_tmp;
 
-    public SM_SO_Service_Exec_Task_Sql_005(long customer_code, int so_prefix, int so_code, int price_list_code, int pack_code, int pack_seq, int category_price_code, int service_code, int service_seq, int exec_tmp, int task_tmp) {
+    public SM_SO_Service_Exec_Task_Sql_005(long customer_code, int so_prefix, int so_code, int price_list_code, int pack_code, int pack_seq, int category_price_code, int service_code, int service_seq, long exec_tmp, long task_tmp) {
         this.customer_code = customer_code;
         this.so_prefix = so_prefix;
         this.so_code = so_code;
